+++ filemgr/src/test/java/org/apache/oodt/cas/filemgr/system/TestXmlRpcFileManagerServer.java
@@ -15,86 +15,34 @@
  * limitations under the License.
  */
 
-package org.apache.oodt.cas.filemgr.system.distributed;
+package org.apache.oodt.cas.filemgr.system;
 
-import org.apache.oodt.cas.filemgr.ingest.StdIngester;
-import org.apache.oodt.cas.filemgr.metadata.CoreMetKeys;
 import org.apache.oodt.cas.filemgr.metadata.ProductMetKeys;
 import org.apache.oodt.cas.filemgr.structs.Product;
 import org.apache.oodt.cas.filemgr.structs.exceptions.CatalogException;
-import org.apache.oodt.cas.filemgr.system.TestXmlRpcFileManagerServer;
-import org.apache.oodt.cas.filemgr.system.XmlRpcFileManager;
-import org.apache.oodt.cas.filemgr.system.XmlRpcFileManagerClient;
-import org.apache.oodt.cas.filemgr.system.XmlRpcFileManagerServer;
 import org.apache.oodt.cas.metadata.Metadata;
-import org.apache.oodt.cas.metadata.SerializableMetadata;
-import org.apache.oodt.cas.metadata.util.PathUtils;
-import org.apache.oodt.config.distributed.cli.ConfigPublisher;
-import org.apache.oodt.config.test.AbstractDistributedConfigurationTest;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.net.URL;
 import java.util.Collections;
-
-import static org.apache.oodt.config.Constants.Properties.ENABLE_DISTRIBUTED_CONFIGURATION;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
- * Tests the {@link XmlRpcFileManager} with distributed configuration management enabled. This test will first publish
- * the required configuration to zookeeper through {@link ConfigPublisher} and will start the {@link XmlRpcFileManager}
- * which will first download and store published files locally. Then the correct functionality of {@link
- * XmlRpcFileManager} is tested by using those downloaded configuration files for configuration.
- * <p>
- * This class is adapted from {@link TestXmlRpcFileManagerServer} class
+ * Test harness for the XmlRpcFileManager.
  *
- * @author Imesha Sudasingha
+ * @since OODT-72
  */
-public class TestDistributedXmlRpcFileManager extends AbstractDistributedConfigurationTest {
+public class TestXmlRpcFileManagerServer extends AbstractFileManagerServerTest {
 
-    private static final int FM_PORT = 9001;
-    private static final String CONF_PUBLISHER_XML = "distributed/config/config-publisher.xml";
-    private static final String TRANSFER_SERVICE_FACTORY_CLASS = "org.apache.oodt.cas.filemgr.datatransfer.LocalDataTransferFactory";
+    private static Logger LOG = Logger.getLogger(TestXmlRpcFileManagerServer.class.getName());
 
-    private XmlRpcFileManagerServer fileManager;
-
-    @Before
-    public void setUpTest() throws Exception {
-        System.setProperty("org.apache.oodt.cas.cli.action.spring.config", "../config/src/main/resources/cmd-line-actions.xml");
-        System.setProperty("org.apache.oodt.cas.cli.option.spring.config", "../config/src/main/resources/cmd-line-options.xml");
-        System.setProperty(ENABLE_DISTRIBUTED_CONFIGURATION, "true");
-        System.setProperty("filemgr.client", "org.apache.oodt.cas.filemgr.system.rpc.XmlRpcFileManagerClientFactory");
-        System.setProperty("filemgr.server", "org.apache.oodt.cas.filemgr.system.rpc.XmlRpcFileManagerServerFactory");
-
-        ConfigPublisher.main(new String[]{
-                "-connectString", zookeeper.getConnectString(),
-                "-config", CONF_PUBLISHER_XML,
-                "-a", "publish"
-        });
-
-        try {
-            fileManager = new XmlRpcFileManagerServer(FM_PORT);
-            if(!fileManager.startUp()){
-              throw new Exception("File Manager startup failed.");
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            fail(e.getMessage());
-        }
-
-        ingestFile();
-    }
-
-    @Test
-    public void testDistributedConfigurationWithFileManager() {
-        XmlRpcFileManagerClient fmc = null;
+    /**
+     * @since OODT-72
+     */
+    public void testExpandProductMet() {
+        FileManagerClient fmc = null;
         try {
             fmc = new XmlRpcFileManagerClient(new URL("http://localhost:" + FM_PORT));
         } catch (Exception e) {
@@ -105,6 +53,7 @@ public class TestDistributedXmlRpcFileManager extends AbstractDistributedConfigu
         try {
             met = fmc.getMetadata(fmc.getProductByName("test.txt"));
         } catch (CatalogException e) {
+            LOG.log(Level.SEVERE, e.getMessage());
             fail(e.getMessage());
         }
 
@@ -145,6 +94,7 @@ public class TestDistributedXmlRpcFileManager extends AbstractDistributedConfigu
         try {
             met = fmc.getReducedMetadata(fmc.getProductByName("test.txt"), Collections.EMPTY_LIST);
         } catch (CatalogException e) {
+            LOG.log(Level.SEVERE, e.getMessage());
             fail(e.getMessage());
         }
 
@@ -181,59 +131,19 @@ public class TestDistributedXmlRpcFileManager extends AbstractDistributedConfigu
         assertEquals("text/plain", met.getMetadata(ProductMetKeys.PRODUCT_MIME_TYPES));
     }
 
-    private void ingestFile() {
-        StdIngester ingester = new StdIngester(TRANSFER_SERVICE_FACTORY_CLASS);
-
-        try {
-            URL ingestUrl = this.getClass().getResource("/ingest");
-            URL refUrl = this.getClass().getResource("/ingest/test.txt");
-            URL metUrl = this.getClass().getResource("/ingest/test.txt.met");
-            Metadata prodMet = new SerializableMetadata(new FileInputStream(new File(metUrl.getFile())));
-
-            // now add the right file location
-            prodMet.addMetadata(CoreMetKeys.FILE_LOCATION, new File(ingestUrl.getFile()).getCanonicalPath());
-            prodMet.addMetadata(CoreMetKeys.FILENAME, "test.txt");
-            prodMet.addMetadata(CoreMetKeys.PRODUCT_TYPE, "GenericFile");
-            ingester.ingest(new URL("http://localhost:" + FM_PORT), new File(refUrl.getFile()), prodMet);
-        } catch (Exception e) {
-            e.printStackTrace();
-            fail(e.getMessage());
-        }
+    @Override
+    protected void setProperties() {
+        System.setProperty("filemgr.server", "org.apache.oodt.cas.filemgr.system.rpc.XmlRpcFileManagerServerFactory");
+        System.setProperty("filemgr.client", "org.apache.oodt.cas.filemgr.system.rpc.XmlRpcFileManagerClientFactory");
     }
 
-    @After
-    public void tearDownTest() throws Exception {
-        if (fileManager != null) {
-            fileManager.shutdown();
-        }
-
-        ConfigPublisher.main(new String[]{
-                "-connectString", zookeeper.getConnectString(),
-                "-config", CONF_PUBLISHER_XML,
-                "-a", "clear"
-        });
-
-        String luceneIdx = System.getProperty("org.apache.oodt.cas.filemgr.catalog.lucene.idxPath");
-        if (luceneIdx != null) {
-            luceneIdx = PathUtils.replaceEnvVariables(luceneIdx);
-            deleteAllFiles(luceneIdx);
-        }
-
-        System.clearProperty("org.apache.oodt.cas.cli.action.spring.config");
-        System.clearProperty("org.apache.oodt.cas.cli.option.spring.config");
-        System.clearProperty(ENABLE_DISTRIBUTED_CONFIGURATION);
+    @Override
+    protected FileManagerServer newFileManagerServer(int port) throws Exception {
+        return new XmlRpcFileManagerServer(port);
     }
 
-    private void deleteAllFiles(String startDir) {
-        File startDirFile = new File(startDir);
-        File[] delFiles = startDirFile.listFiles();
-
-        if (delFiles != null && delFiles.length > 0) {
-            for (File delFile : delFiles) {
-                delFile.delete();
-            }
-        }
-
-        startDirFile.delete();
+    @Override
+    protected boolean shouldExpandProduct() {
+        return true;
     }
 }
